# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de recolección de residuos",
        "title": "API Recolecta",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/anomalias/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anomalia"
                ],
                "summary": "Anomalía por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.AnomaliaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anomalia"
                ],
                "summary": "Actualizar anomalía",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.UpdateAnomaliaRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.AnomaliaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anomalia"
                ],
                "summary": "Eliminar anomalía",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.AnomaliaMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/camion/placa/{placa}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camion"
                ],
                "summary": "Buscar camión por placa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Placa",
                        "name": "placa",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.EstadoCamionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/relleno-sanitario/exists/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RellenoSanitario"
                ],
                "summary": "Verificar existencia relleno sanitario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.RellenoSanitarioListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/rutas/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ruta"
                ],
                "summary": "Listar rutas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.EstadoCamionListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ruta"
                ],
                "summary": "Crear ruta",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.CreateRutaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.RutaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "core.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/core.ErrorDetail"
                }
            }
        },
...
        "entities.Anomalia": {
            "type": "object",
            "properties": {
                "accion_sugerida": {
                    "type": "string"
                },
                "anomalia_id": {
                    "type": "integer"
                },
                "anomalia_referencia_id": {
                    "type": "integer"
                },
                "camion_id": {
                    "type": "integer"
                },
                "categoria_clasificada": {
                    "type": "string"
                },
                "conductor_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "eliminado": {
                    "type": "boolean"
                },
                "estado": {
                    "type": "string"
                },
                "estado_pipeline": {
                    "description": "Pipeline modelo_reportes -\u003e clasificador_reportes. Se llenan por\nActualizarPipeline en background, no al crear la anomalia (por eso\nEstadoPipeline arranca en \"pendiente\" via DEFAULT de la columna).",
                    "type": "string"
                },
                "fecha_reporte": {
                    "type": "string"
                },
                "fecha_resolucion": {
                    "type": "string"
                },
                "inferencia_id": {
                    "type": "integer"
                },
                "json_ruta": {
                    "type": "string"
                },
                "nivel_riesgo": {
                    "type": "string"
                },
                "pipeline_error": {
                    "type": "string"
                },
                "punto_id": {
                    "type": "integer"
                },
                "ruta_id": {
                    "type": "integer"
                },
                "subtipo_clasificado": {
                    "type": "string"
                },
                "tipo_anomalia": {
                    "$ref": "#/definitions/entities.TipoAnomalia"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "entities.AnomaliaMessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "entities.AnomaliaResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/entities.Anomalia"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.CreateRutaRequest": {
            "type": "object",
            "required": [
                "json_ruta",
                "nombre"
            ],
            "properties": {
                "descripcion": {
                    "type": "string"
                },
                "json_ruta": {},
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "entities.EstadoCamion": {
            "type": "object",
            "properties": {
                "camion_id": {
                    "type": "integer"
                },
                "estado": {
                    "description": "nullable",
                    "type": "string"
                },
                "estado_id": {
                    "type": "integer"
                },
                "observaciones": {
                    "description": "nullable",
                    "type": "string"
                },
                "timestamp": {
                    "description": "nullable",
                    "type": "string"
                }
            }
        },
        "entities.EstadoCamionListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.EstadoCamion"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.EstadoCamionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/entities.EstadoCamion"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.RellenoSanitario": {
            "type": "object",
            "properties": {
                "capacidad_toneladas": {
                    "type": "number"
                },
                "direccion": {
                    "type": "string"
                },
                "eliminado": {
                    "type": "boolean"
                },
                "es_rentado": {
                    "type": "boolean"
                },
                "nombre": {
                    "type": "string"
                },
                "relleno_id": {
                    "type": "integer"
                }
            }
        },
        "entities.RellenoSanitarioListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.RellenoSanitario"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.Ruta": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "eliminado": {
                    "type": "boolean"
                },
                "json_ruta": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "ruta_id": {
                    "type": "integer"
                }
            }
        },
        "entities.RutaResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entities.Ruta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.TipoAnomalia": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "TipoAnomaliaAnomalia",
                "TipoAnomaliaIncidencia",
                "TipoAnomaliaReporteConductor",
                "TipoAnomaliaReporteFallaCritica",
                "TipoAnomaliaSeguimientoFallaCritica"
            ]
        },
        "entities.UpdateAnomaliaRequest": {
            "type": "object",
            "required": [
                "descripcion",
                "fecha_reporte",
                "tipo_anomalia"
            ],
            "properties": {
                "anomalia_referencia_id": {
                    "type": "integer"
                },
                "camion_id": {
                    "type": "integer"
                },
                "conductor_id": {
                    "type": "integer"
                },
                "descripcion": {
                    "type": "string"
                },
                "eliminado": {
                    "type": "boolean"
                },
                "estado": {
                    "type": "string"
                },
                "fecha_reporte": {
                    "type": "string",
                    "example": "2026-07-22T19:30:00Z"
                },
                "fecha_resolucion": {
                    "type": "string"
                },
                "json_ruta": {
                    "type": "string"
                },
                "punto_id": {
                    "type": "integer"
                },
                "ruta_id": {
                    "type": "integer"
                },
                "tipo_anomalia": {
                    "type": "string",
                    "example": "REPORTE_CONDUCTOR"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe 'Bearer ' seguido de tu token JWT. Ejemplo: \"Bearer eyJhbG...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}